src/requirements.txt
fastapi==0.104.1
mangum==0.17.0
boto3==1.34.0
botocore==1.34.0
pyjwt==2.8.0
yt-dlp==2023.12.30
python-multipart==0.0.6
uvicorn==0.24.0